Add files via upload
Working on the Julia code to implement the Q-Learning paper. 
Pretty new stuff for me, so the code is very noob level and maybe not very useful. 
For now I have implemented Algo 1 that is given in the paper. Will try to integrate more stuff.
Lmk if you have any suggestions or something specific you want me to work on